Header Position and gallery page
Header was aligned center with all other buttons etc making the title in center which was not good for optics. Updated all html and css files to align it left with logo and changed logo to a more interactive one. Added images from webapp to gallery page using online links