Updates to design and login function
Added condition to display login button on home pages if user is in session and logout button if no user is in session. Used Jinja2 template library for this function.
Added menu bars to remove clutter from pages. Updated all css and html files to implement this function.
Fixed login required condition to all new pages added to the home pages in this menu. For this updated roues.py